feat: searchBar added in User section
