+++ src/views/Admin - Dashboard/pages/Usuarios/Lista-Usuarios.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { getAllUsers } from '../../../../state/features/users/userSlice';
 import DefaultLayout from '../../layout/DefaultLayout';
@@ -9,6 +9,8 @@ import { Pagination } from '../../../../components/Pagination';
 import { titlesUsers } from '../../constants/titleUsers';
 import { Filters } from '../../components/Filters';
 import { filtersUser } from '../../constants/filtersUser';
+import { Search } from '../../../../components/Search';
+import { Result } from '../../components/Result';
 
 const INITIAL_STATE = {
   active: '',
@@ -24,6 +26,8 @@ export const ListaUsuarios = () => {
   const users = useSelector((state) => state.users);
   const filters = useSelector((state) => state.users.filters);
   const pagination = useSelector((state) => state.users.pagination);
+  const [value, setValue] = useState('');
+  const searchBy = 'Buscar por id / correo / nombre';
 
   const handleFilter = (e, type) => {
     if (!e) {
@@ -46,6 +50,17 @@ export const ListaUsuarios = () => {
     }));
   };
 
+  const handleChange = ({ target }) => {
+    setValue(target.value);
+  };
+
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    if (value.trim().length <= 0) return;
+    dispatch(getAllUsers({ ...filters, search: value, currentPage: 1 }));
+    setValue('');
+  };
+
   useEffect(() => {
     dispatch(getAllUsers(filters));
   }, [dispatch]);
@@ -53,6 +68,10 @@ export const ListaUsuarios = () => {
   return (
     <DefaultLayout>
       <Breadcrumb pageName='Lista de Usuarios' />
+      <div>
+        <Search filters={filters} handleChange={handleChange} handleSubmit={handleSubmit} searchBy={searchBy} />
+        <Result filters={filters} handleFilter={handleFilter} />
+      </div>
       <Filters filtersValues={filtersUser} handleFilter={handleFilter} filters={filters} />
       <RowTitles titles={titlesUsers} />
       <RowUsers info={users.list} />
